--- src/SSD1306.c
@@ -1,73 +1,0 @@
-/*
- * SSD1306.c
- *
- *  Created on: Sep 22, 2024
- *      Author: kumar
- */
-
-#include "SSD1306.h"
-#include "stm32f1xx_hal.h"
-
-extern I2C_HandleTypeDef hi2c1;
-
-void SSD1306_I2C_SEND(uint8 control_byte, uint8 data_byte)
-{
-	uint8 sendBuffer[2] = {(uint8)0, (uint8)0};
-	sendBuffer[0] = control_byte;
-	sendBuffer[1] = data_byte;
-
-	/* HAL Library IIC Transmit API */
-	HAL_I2C_Master_Transmit(&hi2c1, oled_write(SSD1306_ADDR), sendBuffer, 2, 10);
-}
-
-void OLED_INIT(void)
-{
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_MUX_RATIO);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, 0x3F);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_DISPLAY_OFFSET);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, 0x00);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_DISPLAY_START_ROWLINE(0));
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_COM_PIN_HW_CONFIG);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, 0x02);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_CONTRAST_CONTROL);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, 0xEB);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, RESUME_RAM_DISPLAY);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, NORMAL_DISPLAY);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, SET_DIV_RATIO_OSC_FREQ);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, 0x80);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, CHARGE_PUMP_SETTING);
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, ENABLE_CHARGE_PUMP);
-
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, DISPLAY(ON));
-}
-
-void OLED_SET_CURSOR(uint8 x, uint8 y)
-{
-	/* Set X coordinate */
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, PAGE_ADDR_MODE_LOW_COL_ADDR(x));
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, PAGE_ADDR_MODE_HIGH_COL_ADDR(x));
-	/* Set y coordinate */
-	SSD1306_I2C_SEND(SSD1306_CTRL_CMD, PAGE_ADDR_MODE_SET_PAGE_START(y));
-}
-
-void OLED_CLEAR()
-{
-	uint8 common = 0, page = 0;
-
-	for(page = 0; page < 8; page++)
-	{
-		for(common = 0; common < 128 ; common++)
-		{
-			OLED_SET_CURSOR(common, page);
-			SSD1306_I2C_SEND(SSD1306_CTRL_DATA, 0x00);
-		}
-	}
-}
